spec/requirements.txt
mkdocs
mkdocs-material
jinja2
pyyaml
mkdocs-awesome-pages-plugin
